Fix a bug when kleaner killed all pods

+++ main.go
@@ -70,15 +70,16 @@ func main() {
 		}
 
 		for _, pod := range pods.Items {
-			strings.Contains(pod.Status.Reason, "Evicted")
-			log.Info("Pod", pod.Name, "will be deleted")
-			err := clientset.CoreV1().Pods(pod.Namespace).Delete(context.TODO(), pod.Name, metav1.DeleteOptions{})
-			if err != nil {
-				log.Error("Unable to delete pod", pod.Name)
-				log.Debug(err.Error())
-				os.Exit(1)
+			if strings.Contains(pod.Status.Reason, "Evicted") {
+				log.Info("Pod", pod.Name, "will be deleted")
+				err := clientset.CoreV1().Pods(pod.Namespace).Delete(context.TODO(), pod.Name, metav1.DeleteOptions{})
+				if err != nil {
+					log.Error("Unable to delete pod", pod.Name)
+					log.Debug(err.Error())
+					os.Exit(1)
+				}
+				log.Info("Pod", pod.Name, "deleted")
 			}
-			log.Info("Pod", pod.Name, "deleted")
 		}
 	}
 }
